Refactor FileDrop styling and use own icon
