+++ src/theme.styled.ts
@@ -11,8 +11,6 @@ export const Theme = styled.div`
     --menubar-box-shadow: 0px 0px 2px 0px #00000088;
     --menubar-border-radius: 4px;
     --background: linear-gradient(#888888, #696969);
-    --button-height: 30px;
-    --button-width: 30px;
     --button-gap: 0px;
     --button-border-radius: 4px;
     --box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
